+++ lib/aruba.rb
@@ -1,8 +1,8 @@
-require 'aruba/spawn_process'
+require 'aruba/processes/spawn_process'
 
 module Aruba
   class << self
     attr_accessor :process
   end
-  self.process = Aruba::SpawnProcess
+  self.process = Aruba::Processes::SpawnProcess
 end
